refactor(rust/devtools): tweak namings and introduction comments (#7028)
Part of https://github.com/rolldown/rolldown/issues/6930.
